Added tests for insertFreeFeedback
